requirements.txt
langchain_community
python-dotenv
streamlit
tiktoken
faiss-cpu
protobuf
